+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/model/KafkaConnectCluster.java
@@ -4,15 +4,22 @@
  */
 package io.strimzi.operator.cluster.model;
 
+import io.fabric8.kubernetes.api.model.ConfigMapVolumeSource;
 import io.fabric8.kubernetes.api.model.Container;
 import io.fabric8.kubernetes.api.model.ContainerBuilder;
 import io.fabric8.kubernetes.api.model.ContainerPort;
 import io.fabric8.kubernetes.api.model.EnvVar;
+import io.fabric8.kubernetes.api.model.EnvVarBuilder;
+import io.fabric8.kubernetes.api.model.EnvVarSource;
+import io.fabric8.kubernetes.api.model.EnvVarSourceBuilder;
 import io.fabric8.kubernetes.api.model.IntOrString;
+import io.fabric8.kubernetes.api.model.SecretVolumeSource;
 import io.fabric8.kubernetes.api.model.Service;
 import io.fabric8.kubernetes.api.model.ServicePort;
 import io.fabric8.kubernetes.api.model.Volume;
+import io.fabric8.kubernetes.api.model.VolumeBuilder;
 import io.fabric8.kubernetes.api.model.VolumeMount;
+import io.fabric8.kubernetes.api.model.VolumeMountBuilder;
 import io.fabric8.kubernetes.api.model.extensions.Deployment;
 import io.fabric8.kubernetes.api.model.extensions.DeploymentStrategy;
 import io.fabric8.kubernetes.api.model.extensions.DeploymentStrategyBuilder;
@@ -25,6 +32,10 @@ import io.strimzi.api.kafka.model.KafkaConnectAuthenticationTls;
 import io.strimzi.api.kafka.model.KafkaConnectS2ISpec;
 import io.strimzi.api.kafka.model.KafkaConnectSpec;
 import io.strimzi.api.kafka.model.PasswordSecretSource;
+import io.strimzi.api.kafka.model.connect.ExternalConfiguration;
+import io.strimzi.api.kafka.model.connect.ExternalConfigurationEnv;
+import io.strimzi.api.kafka.model.connect.ExternalConfigurationEnvVarSource;
+import io.strimzi.api.kafka.model.connect.ExternalConfigurationVolumeSource;
 import io.strimzi.api.kafka.model.template.KafkaConnectTemplate;
 import io.strimzi.operator.common.model.Labels;
 
@@ -47,6 +58,8 @@ public class KafkaConnectCluster extends AbstractModel {
     private static final String METRICS_AND_LOG_CONFIG_SUFFIX = NAME_SUFFIX + "-config";
     protected static final String TLS_CERTS_BASE_VOLUME_MOUNT = "/opt/kafka/connect-certs/";
     protected static final String PASSWORD_VOLUME_MOUNT = "/opt/kafka/connect-password/";
+    protected static final String EXTERNAL_CONFIGURATION_VOLUME_MOUNT_BASE_PATH = "/opt/kafka/external-configuration/";
+    protected static final String EXTERNAL_CONFIGURATION_VOLUME_NAME_PREFIX = "ext-conf-";
 
     // Configuration defaults
     protected static final int DEFAULT_REPLICAS = 3;
@@ -65,6 +78,8 @@ public class KafkaConnectCluster extends AbstractModel {
     protected static final String ENV_VAR_KAFKA_CONNECT_SASL_USERNAME = "KAFKA_CONNECT_SASL_USERNAME";
 
     protected String bootstrapServers;
+    protected List<ExternalConfigurationEnv> externalEnvs = Collections.EMPTY_LIST;
+    protected List<ExternalConfigurationVolumeSource> externalVolumes = Collections.EMPTY_LIST;
 
     private List<CertSecretSource> trustedCertificates;
     private CertAndKeySecretSource tlsAuthCertAndKey;
@@ -205,6 +220,18 @@ public class KafkaConnectCluster extends AbstractModel {
                     kafkaConnect.templateServiceAnnotations = template.getApiService().getMetadata().getAnnotations();
                 }
             }
+
+            if (spec.getExternalConfiguration() != null)    {
+                ExternalConfiguration externalConfiguration = spec.getExternalConfiguration();
+
+                if (externalConfiguration.getEnv() != null && !externalConfiguration.getEnv().isEmpty())    {
+                    kafkaConnect.externalEnvs = externalConfiguration.getEnv();
+                }
+
+                if (externalConfiguration.getVolumes() != null && !externalConfiguration.getVolumes().isEmpty())    {
+                    kafkaConnect.externalVolumes = externalConfiguration.getVolumes();
+                }
+            }
         }
 
         return kafkaConnect;
@@ -250,6 +277,51 @@ public class KafkaConnectCluster extends AbstractModel {
             volumeList.add(createSecretVolume(passwordSecret.getSecretName(), passwordSecret.getSecretName(), isOpenShift));
         }
 
+        volumeList.addAll(getExternalConfigurationVolumes(isOpenShift));
+
+        return volumeList;
+    }
+
+    private List<Volume> getExternalConfigurationVolumes(boolean isOpenShift)  {
+        int mode = 0444;
+        if (isOpenShift) {
+            mode = 0440;
+        }
+
+        List<Volume> volumeList = new ArrayList<>(0);
+
+        for (ExternalConfigurationVolumeSource volume : externalVolumes)    {
+            String name = volume.getName();
+
+            if (name != null) {
+                if (volume.getConfigMap() != null && volume.getSecret() != null) {
+                    log.warn("Volume {} with external Kafka Connect configuration has to contain exactly one volume source reference to either ConfigMap or Secret", name);
+                } else  {
+                    if (volume.getConfigMap() != null) {
+                        ConfigMapVolumeSource source = volume.getConfigMap();
+                        source.setDefaultMode(mode);
+
+                        Volume newVol = new VolumeBuilder()
+                                .withName(EXTERNAL_CONFIGURATION_VOLUME_NAME_PREFIX + name)
+                                .withConfigMap(source)
+                                .build();
+
+                        volumeList.add(newVol);
+                    } else if (volume.getSecret() != null)    {
+                        SecretVolumeSource source = volume.getSecret();
+                        source.setDefaultMode(mode);
+
+                        Volume newVol = new VolumeBuilder()
+                                .withName(EXTERNAL_CONFIGURATION_VOLUME_NAME_PREFIX + name)
+                                .withSecret(source)
+                                .build();
+
+                        volumeList.add(newVol);
+                    }
+                }
+            }
+        }
+
         return volumeList;
     }
 
@@ -274,6 +346,31 @@ public class KafkaConnectCluster extends AbstractModel {
         } else if (passwordSecret != null)  {
             volumeMountList.add(createVolumeMount(passwordSecret.getSecretName(),
                     PASSWORD_VOLUME_MOUNT + passwordSecret.getSecretName()));
+        }
+
+        volumeMountList.addAll(getExternalConfigurationVolumeMounts());
+
+        return volumeMountList;
+    }
+
+    private List<VolumeMount> getExternalConfigurationVolumeMounts()    {
+        List<VolumeMount> volumeMountList = new ArrayList<>(0);
+
+        for (ExternalConfigurationVolumeSource volume : externalVolumes)    {
+            String name = volume.getName();
+
+            if (name != null)   {
+                if (volume.getConfigMap() != null && volume.getSecret() != null) {
+                    log.warn("Volume {} with external Kafka Connect configuration has to contain exactly one volume source reference to either ConfigMap or Secret", name);
+                } else  if (volume.getConfigMap() != null || volume.getSecret() != null) {
+                    VolumeMount volumeMount = new VolumeMountBuilder()
+                            .withName(EXTERNAL_CONFIGURATION_VOLUME_NAME_PREFIX + name)
+                            .withMountPath(EXTERNAL_CONFIGURATION_VOLUME_MOUNT_BASE_PATH + name)
+                            .build();
+
+                    volumeMountList.add(volumeMount);
+                }
+            }
         }
 
         return volumeMountList;
@@ -349,6 +446,44 @@ public class KafkaConnectCluster extends AbstractModel {
             varList.add(buildEnvVar(ENV_VAR_KAFKA_CONNECT_SASL_USERNAME, username));
             varList.add(buildEnvVar(ENV_VAR_KAFKA_CONNECT_SASL_PASSWORD_FILE,
                     String.format("%s/%s", passwordSecret.getSecretName(), passwordSecret.getPassword())));
+        }
+
+        varList.addAll(getExternalConfigurationEnvVars());
+
+        return varList;
+    }
+
+    private List<EnvVar> getExternalConfigurationEnvVars()   {
+        List<EnvVar> varList = new ArrayList<>();
+
+        for (ExternalConfigurationEnv var : externalEnvs)    {
+            String name = var.getName();
+
+            if (name != null && !name.startsWith("KAFKA_") && !name.startsWith("STRIMZI_")) {
+                ExternalConfigurationEnvVarSource valueFrom = var.getValueFrom();
+
+                if (valueFrom != null)  {
+                    if (valueFrom.getConfigMapKeyRef() != null && valueFrom.getSecretKeyRef() != null) {
+                        log.warn("Environment variable {} with external Kafka Connect configuration has to contain exactly one reference to either ConfigMap or Secret", name);
+                    } else {
+                        if (valueFrom.getConfigMapKeyRef() != null) {
+                            EnvVarSource envVarSource = new EnvVarSourceBuilder()
+                                    .withConfigMapKeyRef(var.getValueFrom().getConfigMapKeyRef())
+                                    .build();
+
+                            varList.add(new EnvVarBuilder().withName(name).withValueFrom(envVarSource).build());
+                        } else if (valueFrom.getSecretKeyRef() != null)    {
+                            EnvVarSource envVarSource = new EnvVarSourceBuilder()
+                                    .withSecretKeyRef(var.getValueFrom().getSecretKeyRef())
+                                    .build();
+
+                            varList.add(new EnvVarBuilder().withName(name).withValueFrom(envVarSource).build());
+                        }
+                    }
+                }
+            } else {
+                log.warn("Name of an environment variable with external Kafka Connect configuration cannot start with `KAFKA_` or `STRIMZI`.");
+            }
         }
 
         return varList;